Consolidate tests to avoid unnecessary checkpointing when deleting databases
