+++ libs/checkpoint-postgres/vitest.config.js
@@ -21,7 +21,6 @@ export default defineConfig((env) => {
         exclude: configDefaults.exclude,
         include: ["**/*.int.test.ts"],
         name: "int",
-        fileParallelism: false,
         environment: "node",
       },
     };
